refactor : 모듈 분리
print, getValue 모듈 분리
