+++ src/App.js
@@ -1,5 +1,6 @@
 import { Console } from "@woowacourse/mission-utils";
-import { getAnswer, getUserInput } from "./getRandomNumber.js";
+import { getAnswer, getUserInput } from "./getValue.js";
+import { printStart, printEnd, printResult } from "./printMessage.js";
 
 const SIZE = 3;
 
@@ -48,23 +49,6 @@ const isReStart = async () => {
     return;
   }
   throw new Error("[ERROR]: Invalid input. Please enter 1 or 2");
-};
-
-const printStart = () => Console.print("숫자 야구 게임을 시작합니다.");
-
-const printEnd = () =>
-  Console.print("3개의 숫자를 모두 맞히셨습니다! 게임 종료");
-
-const printResult = (ball, strike) => {
-  if (ball === 0 && strike === 0) {
-    Console.print("낫싱");
-  } else if (ball === 0) {
-    Console.print(`${strike}스트라이크`);
-  } else if (strike === 0) {
-    Console.print(`${ball}볼`);
-  } else {
-    Console.print(`${ball}볼 ${strike}스트라이크`);
-  }
 };
 
 class App {
