+++ src/getValue.js
